pyproject.toml
[tool.poetry]
name = "mdiss"
version = "2.0.1"
description = "Automatyczne generowanie ticketów GitHub na podstawie plików markdown z błędami poleceń"
authors = ["Tom Sapletta <info@softreck.dev>"]
maintainers = ["WRONAI Team"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/wronai/mdiss"
documentation = "https://wronai.github.io/mdiss"
repository = "https://github.com/wronai/mdiss.git"
keywords = ["github", "issues", "automation", "markdown", "ci", "build"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Bug Tracking",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
packages = [{include = "mdiss"}]
include = [
    "LICENSE",
    "README.md",
    "CHANGELOG.md",
    "CONTRIBUTING.md",
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
click = "^8.1.7"
rich = "^13.7.0"
pydantic = "^2.5.0"
domd = "^2.1.4"
python-dotenv = "^1.0.0"
marko = "^2.1.0"
python-gitlab = "^4.0.0"
pygithub = "^2.1.1"
ollama = "^0.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
responses = "^0.24.1"
types-requests = "^2.31.0"
types-python-dateutil = "^2.8.19"
types-pyyaml = "^6.0.12"

[tool.poetry.group.ai]
optional = true

[tool.poetry.group.ai.dependencies]
ollama = "^0.1.5"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.0"
mkdocs-material = "^9.0.0"
mkdocstrings = "^0.20.0"

[tool.poetry.scripts]
mdiss = "mdiss.cli:main"
mdiss-gen = "mdiss.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=mdiss",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["mdiss"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
